fixed videoControllers

+++ controllers/Video/DeleteVideo.ts
@@ -3,22 +3,33 @@ import AsyncHandler from '../../middleware/AsyncHandler';
 import Video from '../../models/Video/Video';
 import ErrorHandler from '../../middleware/ErrorHandler';
 import User from '../../models/User/User';
+import { getUserId } from '../../constants/UserId';
 
 export default AsyncHandler(
   async (req: Request, res: Response, next: NextFunction) => {
     const { videoId } = req.body;
+
+    const userId = (await getUserId(req)).toString();
+
     let video = await Video.findById(videoId);
+
     if (!video)
       return next(new ErrorHandler(404, `Video With Id ${videoId} Not Exist`));
 
-    let user = await User.findById(req['authorizedUser']._id);
+    if (video.user.toString() !== userId)
+      return res.status(404).json({
+        success: false,
+        message: "Sorry!!, You're Not The Owner Of This Video",
+      });
 
-    const videoIndex = user.videos.findIndex(
-      (video) => video['_id'].toString() === videoId
+    let user = await User.findById(userId);
+
+    const videoIndex = user.videos.published.findIndex(
+      (video) => video.toString() === videoId
     );
 
     if (videoIndex >= 0) {
-      user.videos.splice(videoIndex, 1);
+      user.videos.published.splice(videoIndex, 1);
       await user.save();
       await Video.findByIdAndRemove(videoId);
 
@@ -26,10 +37,5 @@ export default AsyncHandler(
         .status(200)
         .json({ success: true, msg: 'Video Deleted Successfully' });
     }
-
-    return res.status(404).json({
-      success: false,
-      message: "Sorry!!, You're Not The Owner Of This Video",
-    });
   }
 );
